fix(Datastore): migrate samples to individual files
